feat(storybook-plugin): move utils in own folder
