+++ packages/storybook-plugin/src/lib/storybook-plugin.ts
@@ -1,29 +1,9 @@
 import fs from 'fs';
-import os from 'os';
-import path from 'path';
 import VirtualModulesPlugin from 'webpack-virtual-modules';
 import { container, Configuration } from 'webpack';
 import { ModuleFederationPluginOptions } from '@module-federation/utilities';
 import { logger } from '@storybook/node-logger';
-
-const correctImportPath = (context: string, entryFile: string) => {
-  if (os.platform() !== 'win32') {
-    return entryFile;
-  }
-
-  if (entryFile.match(/^\.?\.\\/) || !entryFile.match(/^[A-Z]:\\\\/i)) {
-    return entryFile.replace(/\\/g, '/');
-  }
-
-  const joint = path.win32.relative(context, entryFile);
-  const relative = joint.replace(/\\/g, '/');
-
-  if (relative.includes('node_modules/')) {
-    return relative.split('node_modules/')[1];
-  }
-
-  return `./${relative}`;
-};
+import { correctImportPath } from '../utils';
 
 const { ModuleFederationPlugin } = container;
